backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.4
pydantic-core==2.18.4
sqlmodel==0.0.19
sqlalchemy==2.0.30
pyjwt==2.8.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
